requirement.txt
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
Werkzeug==3.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.0
tensorflow
tf-keras
protobuf
